requirements.txt
flask>=2.0.0
flask-sqlalchemy>=3.0.0
flask-migrate>=4.0.0
flask-login>=0.6.0
flask-mail>=0.9.1
python-dotenv>=0.19.0
blinker>=1.4
flask-wtf>=1.1.0
wtforms>=3.0.0
email-validator>=1.3.0
PyJWT>=2.4.0
flask-bootstrap>=3.3.7
Pillow>=10.0.0
